[refactor] delete ManualHandlerMapping in SlippWebApplicationInitializer
