+++ nextstep-mvc/src/main/java/nextstep/mvc/exception/UnsupportedHandlerTypeException.java
@@ -1,4 +1,4 @@
-package nextstep.mvc;
+package nextstep.mvc.exception;
 
 public class UnsupportedHandlerTypeException extends RuntimeException {
     private static final String UNSUPPORTED_HANDLER_TYPE_MESSAGE = "지원하지 않는 핸들러 타입입니다.";
